apps/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# HTTP client for external API calls
httpx==0.25.2
aiohttp==3.9.1

# JSON handling
orjson==3.9.10

# Date and time handling
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
websocket-client==1.6.4

# Code quality (development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Development tools
watchfiles==0.21.0

# Self-learning system dependencies
scikit-learn>=1.3.0  # Machine learning models (severity prediction)
Pillow>=10.0.0       # Image processing
imagehash>=4.3.1     # Image similarity detection (deduplication)
pandas>=2.0.0        # Data processing
numpy>=1.24.0        # Numerical computation
schedule>=1.2.0      # Task scheduling for automated cleaning
